# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Outreach System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type LeadStage = (typeof LeadStage)[keyof typeof LeadStage];

export const LeadStage = {
  raw: "raw",
  enriched: "enriched",
  email_generated: "email_generated",
  contacted: "contacted",
  replied: "replied",
  converted: "converted",
  unsubscribed: "unsubscribed",
} as const;

export interface Lead {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  industry?: string | null;
  /** @nullable */
  companySize?: string | null;
  /** @nullable */
  location?: string | null;
  stage: LeadStage;
  /** @nullable */
  emailValid?: boolean | null;
  /** @nullable */
  emailValidationReason?: string | null;
  /** @nullable */
  isHiring?: boolean | null;
  /** @nullable */
  intentSignal?: string | null;
  /** @nullable */
  websiteSummary?: string | null;
  /** @nullable */
  websiteKeywords?: string | null;
  /** @nullable */
  painPoint?: string | null;
  /** @nullable */
  lcapCompliant?: boolean | null;
  /** @nullable */
  lcapReason?: string | null;
  unsubscribed: boolean;
  /** @nullable */
  unsubscribedAt?: string | null;
  /** @nullable */
  lastContactedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  campaignId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateLeadBody {
  firstName: string;
  lastName: string;
  email: string;
  company: string;
  jobTitle: string;
  website?: string;
  linkedinUrl?: string;
  phone?: string;
  industry?: string;
  companySize?: string;
  location?: string;
  campaignId?: number;
  notes?: string;
}

export interface UpdateLeadBody {
  firstName?: string;
  lastName?: string;
  email?: string;
  company?: string;
  jobTitle?: string;
  website?: string;
  linkedinUrl?: string;
  phone?: string;
  industry?: string;
  companySize?: string;
  location?: string;
  stage?: string;
  campaignId?: number;
  notes?: string;
}

export interface ImportLeadsBody {
  leads: CreateLeadBody[];
  campaignId?: number;
}

export interface ImportLeadsResult {
  imported: number;
  skipped: number;
  errors: string[];
}

export type CampaignStatus =
  (typeof CampaignStatus)[keyof typeof CampaignStatus];

export const CampaignStatus = {
  draft: "draft",
  active: "active",
  paused: "paused",
  completed: "completed",
} as const;

export interface Campaign {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  status: CampaignStatus;
  /** @nullable */
  emailSubject?: string | null;
  /** @nullable */
  emailTemplate?: string | null;
  sendingDelayMinutes: number;
  dailyLimit: number;
  totalLeads: number;
  contacted: number;
  replied: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCampaignBody {
  name: string;
  description?: string;
  emailSubject?: string;
  emailTemplate?: string;
  sendingDelayMinutes?: number;
  dailyLimit?: number;
}

export interface UpdateCampaignBody {
  name?: string;
  description?: string;
  status?: string;
  emailSubject?: string;
  emailTemplate?: string;
  sendingDelayMinutes?: number;
  dailyLimit?: number;
}

export type EmailStatus = (typeof EmailStatus)[keyof typeof EmailStatus];

export const EmailStatus = {
  draft: "draft",
  queued: "queued",
  sent: "sent",
  delivered: "delivered",
  opened: "opened",
  replied: "replied",
  bounced: "bounced",
  failed: "failed",
  unsubscribed: "unsubscribed",
} as const;

export interface Email {
  id: number;
  leadId: number;
  /** @nullable */
  campaignId?: number | null;
  subject: string;
  body: string;
  /** @nullable */
  hook?: string | null;
  status: EmailStatus;
  /** @nullable */
  providerMessageId?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  scheduledAt?: string | null;
  /** @nullable */
  sentAt?: string | null;
  /** @nullable */
  openedAt?: string | null;
  /** @nullable */
  repliedAt?: string | null;
  /** @nullable */
  deliveredAt?: string | null;
  /** @nullable */
  bouncedAt?: string | null;
  createdAt: string;
}

export interface GenerateEmailBody {
  leadId: number;
  campaignId?: number;
  customContext?: string;
}

export interface UpdateEmailBody {
  subject?: string;
  body?: string;
  hook?: string;
}

export interface DashboardStats {
  totalLeads: number;
  enrichedLeads: number;
  emailsGenerated: number;
  emailsSent: number;
  emailsOpened: number;
  emailsReplied: number;
  activeCampaigns: number;
  replyRate: number;
  openRate: number;
}

export interface PipelineStage {
  stage: string;
  count: number;
  label: string;
}

export interface SenderSettings {
  id: number;
  senderName: string;
  senderEmail: string;
  senderCompany: string;
  senderAddress: string;
  pocMessage: string;
  valueProposition: string;
  dailyLimit: number;
  delayMinSeconds: number;
  delayMaxSeconds: number;
  resendEnabled: boolean;
  resendConfigured: boolean;
  updatedAt: string;
}

export interface UpdateSenderSettingsBody {
  senderName?: string;
  senderEmail?: string;
  senderCompany?: string;
  senderAddress?: string;
  pocMessage?: string;
  valueProposition?: string;
  dailyLimit?: number;
  delayMinSeconds?: number;
  delayMaxSeconds?: number;
  resendEnabled?: boolean;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  lead_added: "lead_added",
  lead_enriched: "lead_enriched",
  email_generated: "email_generated",
  email_sent: "email_sent",
  email_opened: "email_opened",
  email_replied: "email_replied",
  campaign_started: "campaign_started",
  campaign_paused: "campaign_paused",
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  leadName?: string | null;
  /** @nullable */
  campaignName?: string | null;
  createdAt: string;
}

export type ListLeadsParams = {
  stage?: string;
  /**
   * @nullable
   */
  campaignId?: number | null;
  search?: string;
};

export type ListEmailsParams = {
  /**
   * @nullable
   */
  leadId?: number | null;
  /**
   * @nullable
   */
  campaignId?: number | null;
  status?: string;
};