# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Outreach System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List all campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCampaignBody"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/start:
    post:
      operationId: startCampaign
      tags: [campaigns]
      summary: Start a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}/pause:
    post:
      operationId: pauseCampaign
      tags: [campaigns]
      summary: Pause a campaign
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"

  # ─────────── EMAILS ───────────
  /emails:
    get:
      operationId: listEmails
      tags: [emails]
      summary: List emails
      parameters:
        - in: query
          name: leadId
          schema:
            type: ["integer", "null"]
        - in: query
          name: campaignId
          schema:
            type: ["integer", "null"]
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Email"
  /emails/generate:
    post:
      operationId: generateEmail
      tags: [emails]
      summary: Generate a personalized AI email for a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateEmailBody"
      responses:
        "201":
          description: Generated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}:
    get:
      operationId: getEmail
      tags: [emails]
      summary: Get an email by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
    patch:
      operationId: updateEmail
      tags: [emails]
      summary: Update email content
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmailBody"
      responses:
        "200":
          description: Updated email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"
  /emails/{id}/send:
    post:
      operationId: sendEmail
      tags: [emails]
      summary: Send an email
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Email"

  # ─────────── DASHBOARD ───────────
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, active, paused, completed]
        emailSubject:
          type: ["string", "null"]
        emailTemplate:
          type: ["string", "null"]
        sendingDelayMinutes:
          type: integer
        dailyLimit:
          type: integer
        totalLeads:
          type: integer
        contacted:
          type: integer
        replied:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - status
        - sendingDelayMinutes
        - dailyLimit
        - totalLeads
        - contacted
        - replied
        - createdAt
        - updatedAt
    CreateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        emailSubject:
          type: string
        emailTemplate:
          type: string
        sendingDelayMinutes:
          type: integer
        dailyLimit:
          type: integer
      required:
        - name
    UpdateCampaignBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          type: string
        emailSubject:
          type: string
        emailTemplate:
          type: string
        sendingDelayMinutes:
          type: integer
        dailyLimit:
          type: integer
    Email:
      type: object
      properties:
        id:
          type: integer
        leadId:
          type: integer
        campaignId:
          type: ["integer", "null"]
        subject:
          type: string
        body:
          type: string
        hook:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, queued, sent, delivered, opened, replied, bounced, failed, unsubscribed]
        providerMessageId:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        sentAt:
          type: ["string", "null"]
        openedAt:
          type: ["string", "null"]
        repliedAt:
          type: ["string", "null"]
        deliveredAt:
          type: ["string", "null"]
        bouncedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - leadId
        - subject
        - body
        - status
        - createdAt
    GenerateEmailBody:
      type: object
      properties:
        leadId:
          type: integer
        campaignId:
          type: integer
        customContext:
          type: string
      required:
        - leadId
    UpdateEmailBody:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        hook:
          type: string
...